+++ examples/spa-minimum-static-build/batfish.config.js
@@ -2,6 +2,8 @@ const path = require('path');
 
 module.exports = () => {
   return {
-    webpackStaticStubReactComponent: [path.join(__dirname, 'src/app.js')]
+    stylesheets: [path.join(__dirname, 'src/style.css')],
+    webpackStaticStubReactComponent: [path.join(__dirname, 'src/app.js')],
+    staticHtmlInlineDeferCss: false
   };
 };